Move externalities implementation to in-core
