+++ packages/jam/in-core/refine.ts
@@ -1,7 +1,6 @@
 import {
   type CoreIndex,
   type Segment,
-  type SegmentIndex,
   type ServiceGas,
   type ServiceId,
   tryAsCoreIndex,
@@ -13,27 +12,23 @@ import type { WorkItem, WorkItemExtrinsic } from "@typeberry/block/work-item.js"
 import { WorkExecResult, WorkExecResultKind, WorkRefineLoad, WorkResult } from "@typeberry/block/work-result.js";
 import { BytesBlob } from "@typeberry/bytes";
 import { codec, Encoder } from "@typeberry/codec";
-import type { KnownSizeArray } from "@typeberry/collections";
 import type { ChainSpec, PvmBackend } from "@typeberry/config";
 import { PvmExecutor, type RefineHostCallExternalities, ReturnStatus, type ReturnValue } from "@typeberry/executor";
 import { type Blake2b, HASH_SIZE } from "@typeberry/hash";
 import { tryAsU32 } from "@typeberry/numbers";
 import type { State } from "@typeberry/state";
 import type { WorkPackageFetchData } from "@typeberry/transition/externalities/fetch-externalities.js";
-import { RefineFetchExternalities } from "@typeberry/transition/externalities/refine-fetch-externalities.js";
 import { assertNever, Result } from "@typeberry/utils";
-import { RefineExternalitiesImpl } from "./externalities/refine.js";
+import {
+  type ImportedSegment,
+  type PerWorkItem,
+  RefineExternalitiesImpl,
+  RefineFetchExternalities,
+} from "./externalities/index.js";
 
 export type RefineItemResult = {
   result: WorkResult;
   exports: readonly Segment[];
-};
-
-export type PerWorkItem<T> = KnownSizeArray<T, "for each work item">;
-
-export type ImportedSegment = {
-  index: SegmentIndex;
-  data: Segment;
 };
 
 enum ServiceCodeError {
