+++ packages/jam/transition/externalities/index.ts
@@ -1,5 +1,3 @@
 export * from "./accumulate-externalities.js";
 export * from "./accumulate-fetch-externalities.js";
 export * from "./fetch-externalities.js";
-export * from "./is-authorized-fetch-externalities.js";
-export * from "./refine-fetch-externalities.js";
